# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Trade Sovereign API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /media:
    get:
      operationId: listMedia
      tags: [media]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [music, movie]
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
  /media/{id}:
    get:
      operationId: getMediaItem
      tags: [media]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Media item details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
  /media/{id}/download:
    get:
      operationId: getMediaDownloadUrl
      tags: [media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResponse"
        "403":
          description: Not purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /pages/{slug}:
    get:
      operationId: getPage
      tags: [pages]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Page content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/media:
    get:
      operationId: adminListMedia
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
    post:
      operationId: adminCreateMedia
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaRequest"
      responses:
        "201":
          description: Media item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
    post:
      operationId: adminCreateCategory
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/pages:
    get:
      operationId: adminListPages
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pages list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageListResponse"
    post:
      operationId: adminCreatePage
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageRequest"
      responses:
        "201":
          description: Page created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
  /admin/pages/{id}:
    put:
      operationId: adminUpdatePage
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageRequest"
      responses:
        "200":
          description: Page updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    delete:
      operationId: adminDeletePage
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        productCount:
          type: number
        createdAt:
          type: string
      required: [id, name, slug, createdAt]
    CategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        total:
          type: number
      required: [categories, total]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
      required: [name, slug]
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        salePrice:
          type: number
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        stock:
          type: number
        imageUrl:
          type: string
        isDigital:
          type: boolean
        isSubscription:
          type: boolean
        createdAt:
          type: string
      required: [id, name, price, category, stock, isDigital, isSubscription, createdAt]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required: [products, total]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        salePrice:
          type: number
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        stock:
          type: number
        imageUrl:
          type: string
        isDigital:
          type: boolean
        isSubscription:
          type: boolean
      required: [name, price, category, stock, isDigital, isSubscription]
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        salePrice:
          type: number
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        stock:
          type: number
        imageUrl:
          type: string
        isDigital:
          type: boolean
        isSubscription:
          type: boolean
    MediaItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [music, movie]
        price:
          type: number
        description:
          type: string
        imageUrl:
          type: string
        fileUrl:
          type: string
        licenseType:
          type: string
        createdAt:
          type: string
      required: [id, title, type, price, licenseType, createdAt]
    MediaListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required: [items, total]
    CreateMediaRequest:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [music, movie]
        price:
          type: number
        description:
          type: string
        imageUrl:
          type: string
        fileUrl:
          type: string
        licenseType:
          type: string
      required: [title, type, price, licenseType]
    DownloadUrlResponse:
      type: object
      properties:
        url:
          type: string
        expiresAt:
          type: string
      required: [url, expiresAt]
...
    Page:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        contentType:
          type: string
          enum: [markdown, html]
        isPublished:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, content, contentType, isPublished, createdAt, updatedAt]
    PageListResponse:
      type: object
      properties:
        pages:
          type: array
          items:
            $ref: "#/components/schemas/Page"
        total:
          type: number
      required: [pages, total]
    CreatePageRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        contentType:
          type: string
          enum: [markdown, html]
        isPublished:
          type: boolean
      required: [title, slug, content, contentType, isPublished]
...