# lib/api-zod/src/generated/types/createMediaRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Sovereign API specification
 * OpenAPI spec version: 0.2.0
 */
import type { CreateMediaRequestType } from "./createMediaRequestType";

export interface CreateMediaRequest {
  title: string;
  type: CreateMediaRequestType;
  price: number;
  description?: string;
  imageUrl?: string;
  fileUrl?: string;
  licenseType: string;
}