# lib/api-zod/src/generated/types/mediaListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Sovereign API specification
 * OpenAPI spec version: 0.2.0
 */
import type { MediaItem } from "./mediaItem";

export interface MediaListResponse {
  items: MediaItem[];
  total: number;
  page?: number;
  limit?: number;
}